# Optimize cast(string as double/float) by switching to fast_float library