Fixes: https://github.com/facebookincubator/velox/issues/9119

CastBenchmark(Benchmark running in RELEASE mode):

BEFORE:
```
/Users/wypb/data/code/apache/velox/_build/release/velox/benchmarks/basic/velox_cast_benchmark
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast##cast_string_as_double                                36.49ms     27.40
cast##cast_string_as_float                                 31.05ms     32.20
```
AFTER:
```
/Users/wypb/data/code/apache/velox/_build/release/velox/benchmarks/basic/velox_cast_benchmark
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast##cast_string_as_double                                 3.66ms    272.90
cast##cast_string_as_float                                  3.70ms    270.01

```

CC: @mbasmanova