@kgpai 2024-03-19
Thank you for this contribution @wypb ; As an FYI, we have an older version of fast_float library internally and would need to be upgraded to be compatible with this version , which will require some time and testing.