@kgpai 2024-05-22
Hi @wypb , 

There are some considerations here, 

1. Firstly , we have been trying to get equivalence between Presto java's cast of floating point values to string.  There is a PR here exploring use of double_conversion for this purpose here : https://github.com/facebookincubator/velox/pull/9747 . To help with this there is a cast fuzzer which runs against Presto to check accuracy of the converted floating point value. Feel free to take the cast fuzzer from there and test it against your changes to see how close it is to java's cast. 
2. The reason we want to do this is we are trying to validate Prestissimo against Presto java and we get a lot of verification errors and thus being able to be equivalent to Presto java would be huge. 
3. Finally if fast_float is equivalent it will help drive the work to upgrade it internally since with any upgrade we need to make sure that we dont break any existing functionality esp as this library is used by other teams.