@wypb 2024-05-23
HI @kgpai  Thank you for the information; it was very helpful for understanding the next steps we need to take. 

The performance of `fast_float` in parsing data is indeed great, but from the perspective of data consistency between Presto and Prestissimo, I think `double_conversion` would be more appropriate.

I will modify my internal code based on [#9747](https://github.com/facebookincubator/velox/pull/9747). If there are any performance issues, I will raise an issue to the community.