@mbasmanova 2024-05-23
@kgpai Krishna, my understanding is fast_float can be used to optimize casting strings to float/double, not the other way around. Your concern about matching Presto seems to be re: the opposite conversion (float/double to string), no?