@mbasmanova 2025-11-06
@rui-mo In case @wypb doesn't have bandwidth / interest in advancing this optimization, would you or your teammates be interested in picking it up?

CC: @amitkdutta